Add 90.py starting automatically

+++ LolexTools.py
@@ -25,21 +25,31 @@ except(ImportError):
     else:
         os.system("python3 ./LolexToolsInstaller.py")
     exit(None)
-if LolexToolsOptions.compiledon<8.3:
-    shutil.copy("./update/83.py","./")
+if LolexToolsOptions.compiledon<9.0:
+    if LolexToolsOptions.compiledon<8.3:
+        shutil.copy("./update/83.py","./")
+        if system == "Windows":
+            if sys.version_info.minor>5:
+                os.system ("py ./83.py")
+            else:
+                os.system("python ./83.py")
+        else:
+            os.system ("python3 ./83.py")
+        os.remove("./83.py")
+    shutil.copy("./update/90.py","./")
     if system == "Windows":
         if sys.version_info.minor>5:
-            os.system ("py ./83.py")
+            os.system ("py ./90.py")
         else:
-            os.system("python ./83.py")
+            os.system("python ./90.py")
     else:
-        os.system ("python3 ./83.py")
-    os.remove("./83.py")
+        os.system ("python3 ./90.py")
+    os.remove("./90.py")
     if system == "Windows":
         if sys.version_info.minor>5:
-            os.system("py .\start.py")
+          os.system("py .\start.py")
         else:
-            os.system("python .\start.py")
+           os.system("python .\start.py")
     else:
         os.system("python3 ./start.py")
     exit(None)
